move common logic to shared folder
